feat(agent): agent schema (#2273)
* adds agent config schema

* adjusts agent config schema

* separates context migrate from agent file for readability

* vends agent schema using schemars

* adds agent subcommand to print config schema

* changes hooks types to untagged union in agent config
