+++ Cargo.toml
@@ -126,6 +126,7 @@ whoami = "1.6.0"
 windows = { version = "0.61.1", features = ["Foundation", "Win32_System_ProcessStatus", "Win32_System_Kernel", "Win32_System_Threading", "Wdk_System_Threading"] }
 winnow = "=0.6.2"
 winreg = "0.55.0"
+schemars = "1.0.4"
 
 [workspace.lints.rust]
 future_incompatible = "warn"
