+++ crates/chat-cli/src/cli/chat/tools/custom_tool.rs
@@ -8,6 +8,7 @@ use crossterm::{
     style,
 };
 use eyre::Result;
+use schemars::JsonSchema;
 use serde::{
     Deserialize,
     Serialize,
@@ -37,15 +38,20 @@ use crate::mcp_client::{
 use crate::os::Os;
 
 // TODO: support http transport type
-#[derive(Clone, Serialize, Deserialize, Debug, Eq, PartialEq)]
+#[derive(Clone, Serialize, Deserialize, Debug, Eq, PartialEq, JsonSchema)]
 pub struct CustomToolConfig {
+    /// The command string used to initialize the mcp server
     pub command: String,
+    /// A list of arguments to be used to run the command with
     #[serde(default)]
     pub args: Vec<String>,
+    /// A list of environment variables to run the command with
     #[serde(skip_serializing_if = "Option::is_none")]
     pub env: Option<HashMap<String, String>>,
+    /// Timeout for each mcp request in ms
     #[serde(default = "default_timeout")]
     pub timeout: u64,
+    /// A boolean flag to denote whether or not to load this mcp server
     #[serde(default)]
     pub disabled: bool,
 }
